fix: add refactor to build folder scturcture add template
