+++ cmd/web/main.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"log"
+	"net/http"
+)
+
+func main() {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/", home)
+	mux.HandleFunc("/snippet", showSnippet)
+	mux.HandleFunc("/snippet/create", createSnippet)
+
+	log.Println("Starting server on :8080")
+	err := http.ListenAndServe(":8080", mux)
+	log.Fatal(err)
+
+}
