+++ ui/html/home.page.tmpl
@@ -1,0 +1,19 @@
+<!doctype html>
+<html lang='en'>
+<head>
+    <meta charset='utf-8'>
+    <title>Home - Snippetbox</title>
+</head>
+<body>
+<header>
+    <h1><a href='/'>Snippetbox</a></h1>
+</header>
+<nav>
+    <a href="/">Home</a>
+</nav>
+<main>
+    <h2>Latest Snippets</h2>
+    <p>There's nothing to see here yet!</p>
+</main>
+</body>
+</html>
